IS Project/backend/requirements.txt
# Web framework
fastapi
uvicorn[standard]
pydantic
joblib

# Vector store + RAG
langchain
langchain-community
langchain-pinecone
langchain-openai
pinecone
openai

# LLM API 
volcengine-python-sdk[ark]

# Utils
python-dotenv